function log out
